requirements.txt
streamlit
python-dotenv
google-generative-ai
